Store the attention KV cache as FP16
Keys and values round to half once at write in both the decode and the
prefill kernels, so the two paths' cache states stay bitwise-comparable
and the verification API converts back to float transparently. Memory
per context token drops from 128 KiB to 64 KiB (0.54 GB -> 0.27 GB at
context 4096) and attention read traffic halves as the context grows.

Battery output is token-identical with the MTP adaptive default and
decode rates are unchanged at short context. The parity test gains a
0.8% relative tolerance escape (one half ulp at key magnitudes in the
hundreds is 0.25-0.5, which the flat absolute bound would misread as
drift) and a measured 0.75 bound for the half-tile GEMM mode, which
reaches 0.41 drift with identical argmax decisions once the rounded
cache feeds it. Both regression suites pass.

+++ models/qwen3.6-27b/targets/apple-m3-pro/qwen36_attention.metal
@@ -101,8 +101,8 @@ kernel void qwen36_attention_prepare_key_value(
     device const float *projected [[buffer(0)]],
     device const float *norm_weight [[buffer(1)]],
     constant AttentionDecodeParameters &parameters [[buffer(2)]],
-    device float *key_cache [[buffer(3)]],
-    device float *value_cache [[buffer(4)]],
+    device half *key_cache [[buffer(3)]],
+    device half *value_cache [[buffer(4)]],
     uint tid [[thread_index_in_threadgroup]],
     uint3 group_id [[threadgroup_position_in_grid]]) {
     threadgroup float values[256];
@@ -131,7 +131,7 @@ kernel void qwen36_attention_prepare_key_value(
 
 kernel void qwen36_attention_scores(
     device const float *query [[buffer(0)]],
-    device const float *key_cache [[buffer(1)]],
+    device const half *key_cache [[buffer(1)]],
     constant AttentionDecodeParameters &parameters [[buffer(2)]],
     device float *scores [[buffer(3)]],
     uint lane [[thread_index_in_simdgroup]],
@@ -159,7 +159,7 @@ kernel void qwen36_attention_scores(
 
 kernel void qwen36_attention_softmax_value_gate(
     device const float *scores [[buffer(0)]],
-    device const float *value_cache [[buffer(1)]],
+    device const half *value_cache [[buffer(1)]],
     device const float *query_gate [[buffer(2)]],
     constant AttentionDecodeParameters &parameters [[buffer(3)]],
     device float *output [[buffer(4)]],
